Fix output for "openstack volume backup restore"
Previously, on Openstack Pike, an issue [1] was reported that the
`openstack volume backup restore` command was not able to parse the
output correctly, even though the restore operation succeeds. This was
fixed by [2], for the Stein release (and cherry-picked to both Rocky
and Queens). The issue was that cliff expects a list with two tuples to
display the results, whereas the restore function was returning a
VolumeBackupsRestore object. The solution was to use the `_info` field
from the VolumeBackupsRetore object, instead of the whole object.

This was done not only for the VolumeBackupsRetore object, but also for
the VolumeBackup one, as can be seen on [3] and [4]. However, the
commit [5] removed this essential parsing, and caused the previously
fixed issue to reappear.

[1] https://bugs.launchpad.net/python-openstackclient/+bug/1733315
[2] https://review.opendev.org/c/openstack/python-openstackclient/+/624860
[3] https://opendev.org/openstack/python-openstackclient/src/branch/stable/2023.1/openstackclient/volume/v2/volume_backup.py#L174
[4] https://opendev.org/openstack/python-openstackclient/src/branch/stable/2023.1/openstackclient/volume/v2/volume_backup.py#L619
[5] https://review.opendev.org/c/openstack/python-openstackclient/+/353931

Test Plan:
PASS: Build python-openstackclient package
PASS: Build wheels
PASS: Build stx-openstackclients image
PASS: Manually upload built image to the system's registry and perform
      helm-override to use it in the clients containers
PASS: Create a backup from a volume
PASS: Issue the `openstack volume backup restore` command and verify
      that the output is returned without any errors.

Closes-Bug: 2045183

Change-Id: I18752635d29e0b74b30b0d63b6194c28d947d5a0
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>