Add Neutron DHCP HA capabilities
Aiming for the stx-openstack application hardening, this change
implements the plugin logic to gather the number of available compute
nodes and, for anything that is not an AIO-SX deployment, configure the
neutron dhcp_agents_per_network count accordingly [1].
When only two computes are available this config is set to 2, whenever
three or more compute nodes are available this config is set to 3.

There are scenarios in which the compute node fails and if it is hosting
the DHCP agent for a given openstack network, VMs on that VM won't be
able to receive DHCPOFFER and fail to receive ip and ping.

Partial-Bug: 2030883

[1] https://docs.openstack.org/neutron/latest/admin/config-dhcp-ha.html#enabling-dhcp-high-availability-by-default

TEST PLAN:
PASS - Build plugins (k8sapp_openstack)
PASS - Build stx-openstack helm charts
PASS - Apply stx-openstack and verify the Neutron config
       * Check dhcp_agents_per_network count:
       /etc/neutron/neutron.conf
       - AIO-SX = 1
       - AIO-DX = 2
PASS - Execute stx-openstack sanity tests locally

Change-Id: I6f32ac8a1706467be0be67920a3a750bb7c6b01e
Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>