Add platform label to pods in stx-openstack
Add the label app.starlingx.io/component to the pods and allow the
value to be updated via "system helm-override-update". This change
also ensures that when changing the label value and reapplying the
app, the pod is restarted.

The value of the label can only be "platform" or "application", if
the value is different from that, the label will not be applied.
By default, all pods start with the value platform.

Patches for the openstack-helm and openstack-helm-infra
repositories with the same purpose will be created later.

Test Plan:
PASS: upload/apply/remove/delete/update stx-openstack
PASS: Install stx-openstack and check if pods have the label
'app.starlingx.io/component' with the value 'platform'.
PASS: After updating the isApplication variable to true for
nova-api-proxy chart using "system update-helm-override" and
reapplying stx-openstack, the pods were restarted and the label
was successfully changed.
PASS: If "isApplication" is updated with a value other than
true or false, the label on the pods "app.starlingx.io/component"
will not change.

Story: 2010612
Task: 49294

Change-Id: I984be94a42cc302357d0b2202be1425e80a52505
Signed-off-by: David Barbosa Bastos <david.barbosabastos@windriver.com>