Add platform label to pods in upstream repositories
Add the label app.starlingx.io/component to the pods via patches
for the openstack-helm and openstack-helm-infra repositories and
allow the value to be updated via "system helm-override-update".
This change also ensures that when changing the label value and
reapplying the app, the pod is restarted.

The value of the label can only be "platform" or "application", if
the value is different from that, the label will not be applied.
By default, all pods start with the value platform.

Test Plan:
PASS: build-pkgs -c -p openstack-helm
PASS: build-pkgs -c -p openstack-helm-infra
PASS: build-pkgs -c -p stx-OpenStack-helm-fluxcd
PASS: upload/apply/remove/delete stx-openstack
PASS: Install stx-openstack and check if pods have the label
'app.starlingx.io/component' with the value 'platform'.
PASS: After updating the isApplication variable to true for
ingress chart using "system update-helm-override" and reapplying
stx-openstack, the pods were restarted and the label was
successfully changed.
PASS: After updating the isApplication variable to true for
cinder chart using "system update-helm-override" and reapplying
stx-openstack, the pods were restarted and the label was
successfully changed.
PASS: If "isApplication" is updated with a value other than
true or false, the label on the pods "app.starlingx.io/component"
will not change.

Story: 2010612
Task: 49299

Change-Id: Ifb6e43e15e19e5b17aefd9c5b611d65b7543ded0
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>