Add new label to fm-rest-api
Add the label 'isApplication' to the fm-rest-api jobs and FluxCD
manifest's static-override.

The value of the label can be set to 'true' or 'false'. By changing
this value, the pods will be deployed either with the value of
'application' or 'platform' in the 'app.starlingx.io/component' label.
By default, the new label is set to 'false', thus all pods start with
the value 'platform'.

The label 'ttlSecondsAfterFinished' was also added to the helm-toolkit's
jobs, that are used by fm-rest-api, through a patch. This label was added
so the application can be reapplied, after helm-override, without the
need of removing the application first.

The value of this label was set to 200 seconds

Test Plan:
PASS - The openstack-armada-app is successfully built
PASS - stx-openstack is successfully built
PASS - The pods are deployed with label value set to platform
PASS - The pods value is changing using helm-override-update and reapplying
       the application.

Story: 2010612
Task: 50518

Depends-On: https://review.opendev.org/c/starlingx/helm-charts/+/924177

Change-Id: I1406744aec7c19fab8b5da795d563e5912fad8e8
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>