Expose Job ttl on values and overrides files
STX-Openstack apply was failing due to a low value of a ttl
set to the jobs. Some jobs were being cleaned before the
deployment of the Helm chart was concluded.

This review creates patches to expose the ttl configuration
on the values file of all Helm charts used by STX-Openstack.
The values set in the values and static override files is
relative to 1h30m, the same set as default inside the Job template.

Magnum Helm chart was causing problems during the build, and since
it is not supported by STX-Openstack it was removed from the rules files
for the openstack-helm build.

Test Plan:
PASS - STX-Openstack is built
PASS - Upload and apply STX-O
PASS - Jobs are cleaned after 1h30m
PASS - helm-override-update changes the value of ttl
PASS - Remove and Delete STX-O

Depends-on: https://review.opendev.org/c/starlingx/helm-charts/+/925465

Closes-Bug: 2075509

Change-Id: I11f106875244f1575c8be8f47ca6ca3002320c83
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>