Revert ttl addition to STX-Openstack jobs
The ttl label was added as a way to handle the reapply of
the application after a user override [1][2].

The deletion of the jobs using the ttlSecondsAfterFinished
was causing some pods of the application to be stuck in
init state after a host reboot.

Some pods in the STX-Openstack have job dependencies, meaning that
they will only start if this jobs exist and are completed.

This review removes the ttlSecondsAfterFinished from the jobs in
the application.

[1]: https://review.opendev.org/c/starlingx/openstack-armada-app/+/924351
[2]: https://review.opendev.org/c/starlingx/openstack-armada-app/+/925481

Test Plan:
PASS: STX-Openstack is built
PASS: STX-Openstack upload and apply
PASS: All pods come back online after host reboot
PASS: Remove and Delete STX-Openstack

Partial-Bug: 2076118

Change-Id: I0a4392701b255ea2aeb6bc942d085dc3588ca641
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>