Add pre-apply cleanup Job to STX-O Helm charts
After verification, it was noted that it is not possible
to reapply STX-Openstack after a helm-override that changes
a job template since the template section of a job is
immutable or not updatable

Due to the use of kubernetes-entrypoint DEPENDENCY_JOBS it was
also noted that deleting the jobs after the application is applied
it is not an option. If this happens, the application would not
come back after a host reboot.

This review creates a Job that runs right before the Helm chart is
installed or updated. This Job deletes all jobs that have its status
as completed.

Test Plan:
PASS - Build STX-Openstack
PASS - Upload and Apply STX-O
PASS - Reapply STX-O after a helm-override
PASS - STX-O up after host reboot
PASS - Remove and Delete STX-O

Story: 2010612
Task: 50856

Change-Id: Ia1573af73a1ec23e053ddac127658dcae441e810
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>