Decouple STX-O App Certificates from the Platform
This commit updates the handling of certificates for the STX-Openstack
application by decoupling them from the platform. The primary reasons
for this update are:

1. To facilitate the deprecation of Openstack-related CLI commands
   for certificate management by the platform team [1].
2. To align with the initiative of containerizing and separating the
   STX-Openstack application from the platform.

The modification involves updating the application's plugins to avoid
fetching HTTPS information from the sysinv API. Instead, the plugins now
check for the presence of a certificate file at a specified location
within the OPENSTACK_WORKING_DIR [2].

To support this change, new Helm variables have been introduced in the
clients chart. While these variables are not directly utilized by the
chart, they provide end-users with the ability to customize the path and
name of the certificate file. If a valid certificate file is detected,
HTTPS will be automatically enabled for the application.

Additionally, all functions previously relying on the https_enabled
field in the sysinv database API have been removed, as this field is
now assumed to be always true and cannot be altered by the user [3].

This change also updates all tests cases so that they use the updated
functions.

[1] https://opendev.org/starlingx/config/src/branch/master/sysinv/sysinv/sysinv/sysinv/api/controllers/v1/certificate.py#L267
[2] https://opendev.org/starlingx/openstack-armada-app/src/branch/master/stx-openstack-helm-fluxcd/stx-openstack-helm-fluxcd/helm-charts/clients/values.yaml#L174
[3] https://opendev.org/starlingx/config/commit/732437f3cd9d8ed7170e4870314cbd2e8a569fa4#diff-6ace355d52527e7f52d8f13fddbb559110e31923

Test Plan:
  - PASS: Build STX-Openstack tarball
  - PASS: Apply STX-Openstack
  - PASS: Verify all app communications is being done via HTTP
  - PASS: Add the certificate file to /var/opt/openstack/ssl and reapply
          the app successfully.
  - PASS: Verify all app communications is being done via HTTPS
  - PASS: Modify the default value for the path / name of the
          certificate file and reapply the app. Modify the actual file.
  - PASS: Verify that after the modification, communication is stil
          being done via HTTPS.

Story: 2011208
Task: 50876

Change-Id: Id8601571085e6e2e805f69b2032794363b93d469
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>