Fix tox pep8 job for k8sapp_openstack
The Tox job for pep8 is failing for k8sapp_openstack due to a missing
file in the openstack/requirements repository [1]. This job fetches the
upper constraints from the Stein branch, but it no longer exists.

This commit simply updates the upper constraints [1] branch from
"branch/stable/stein" to "tag/stein-eol". Currently, there's no way to
set this value to a more recent branch (stable/2023.1 for example),
because the packages for these newer versions do not support Python 3.6,
and Tox is failing to use Python 3.9 version for the Pep8 task.

[1] https://opendev.org/openstack/requirements

Test Plan:
  - PASS: Run tox for the python3-k8sapp-openstack folder

Closes-Bug: 2077730

Change-Id: Ic65f1b840098fb048fe9f2ce00d1e33a4a58c557
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>