Add app.kubernetes.io/name label to openstack pods
When trying to apply STX-O in a lab, the application framework reports
that STX-O status is "Applied" even when some pods are not in ready state

This commit modifies the kubernetes_metadata_labels template of the
helm-toolkit adding the app.kubernetes.io/name label. It also removes
the the ones that were added manually to avoid conflicts.

This label allows the application framework to correctly look for the
status of every pod.

Test Plan:
PASS - Validate every that every Openstack pod has the
app.kubernetes.io/name label.
PASS - 'system application-show stx-openstack' only shows applied when
all pods are ready.
PASS - Validate all pods are in the correct state when the status is "Applied"
PASS - Validate the correct status after application re-apply with overrides.

Closes-Bug: 2082019

Change-Id: I5ebdfb445c7fdd979dc2c8aa9cba861c751d8367
Signed-off-by: Johnny Chia <johnny.chialung@windriver.com>