Dynamic FQDN configuration
Currently, the STX-Openstack application only supports FQDN naming as
the following: "service_name.domain_endpoint". This limits how someone
can use the application, as their DNS service might not support this.

As a solution, we can allow the user to change how the services' FQDN
will be configured by using a pattern to create the FQDN. This change
introduces this, allowing one to change the default
"service_name.domain_endpoint" to whatever their DNS supports. For this,
a new entry in the values.yaml file on the Clients chart was added that
contain this pattern. It defaults to "{service_name}.{endpoint_domain}",
and the user can change it according to their infrastructure, using
those 2 fields: service_name and endpoint_domain. The endpoint_domain is
configured by following [1], and the service_name are the various
STX-Openstack services.

[1] https://docs.starlingx.io/security/openstack/update-the-domain-name.html#update-the-domain-name

Test Plan:

- PASS: Build / upload / apply STX-Openstack
- PASS: Apply DNS configuration by following [1] with the template
        "{service_name}.{endpoint_domain}".
- PASS: Apply DNS configuration by following [1] with the template
        "{service_name}-{endpoint_domain}". Apply an override to
        update the pattern in the application.
- PASS: For both scenarios, verify the endpoints were configured by
        using "openstack endpoint list"

Change-Id: Ib6c9c8af574c34d95d6c7ce657d1abd8f81ad811
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>