Add info for installing openstack in a subcloud
Add a condition to check if the deployment is in a subcloud and instruct
the user to create the necessary resources before proceeding. When in an
AIO, a ldap group called "openstack" with gid 1001 is created locally.
When in a subcloud, the group must exist in the system controller, which
is impossible without a major change in how the internal ldap is
configured. So instruct the user to manually create the group.

The check is done by a generic function which can be reused elsewhere.
There are similar functions in the code base but they are not
straightforward to use. They depend on having context, either by being
included as methods in classes and thus dependending on instantiating
objects (even when not needed) or creating context for frameworks
included as dependency (pecan). A simpler function was needed.

Test Plan:

Success path:
    - PASS: Install a system controller and a subcloud
    - PASS: Add an ldap group called "openstack" with gid 1001 in the
            system controller
    - PASS: Install wr-openstack in the subcloud

Failure path:
    - PASS: Install a system controller and a subcloud
    - PASS: Install wr-openstack in the subcloud
    - PASS: Verify an error message related to not being able to create
            a ldap group

Closes-Bug: 2075514

Change-Id: I3665b2b9d7dbdb7a974a10f2ed9b3722fb51c1ae
Signed-off-by: Daniel Bolgheroni <Daniel.Bolgheroni@windriver.com>