Remove auth middleware from nova proxy pipeline
Nova instance actions (e.g., suspend, resume, pause) are failing
intermittently with status HTTP 500. These actions are routed to
the NFVI Compute Plugin running on the platform (NFV/VIM) by the
service nova-api-proxy running as part of the openstack app.

By setting the keystonemiddleware log at Debug level on nova-api-proxy,
we were able to catch the exception "http.client.RemoteDisconnected:
Remote end closed connection without response write" including the
keystonemiddleware in the traceback. To isolate the issue, we have
performed tests by using "curl" command to send nova instance actions
for 3 different scenarios:

- [FAIL] (curl -> nova -> nova-api-proxy -> NFVI Compute Plugin) this
scenario is the same as running the CLI command "openstack server
<action> <instance id>". The request goes through the nova and
nova-api-proxy services until it's proxied to the NFVI Compute Plugin.
As expected, the test failed with error 500;
- [FAIL] (curl -> nova-api-proxy -> NFVI Compute Plugin): this scenario
bypasses the nova service. The request is sent directly to the
nova-api-proxy where it's proxied to the NFVI Compute Plugin. This test
failed with error 500, discarding the possibility that the error is
related to network issues between nova and nova-api-proxy;
- [PASS] (curl -> NFVI compute plugin): this scenario completely
bypasses the openstack app. The request is sent directly to NFVI Compute
Plugin from the host (active controller). This test succeeded, it
doesn't present the error 500, discarding the possibility that the error
is related to NFVI Compute Plugin and confirming that it's related to
the openstack application itself.

Based on these results, we removed the keystone middleware from the
nova-api-proxy pipeline and checked that, after this change, all the
nova instance actions were successful. This seems to be an acceptable
solution since the nova requests are already authenticated in the
openstack endpoints and don't really need to be authenticated by
the nova-api-proxy.

It's still unclear why using the keystone middleware in nova-api-proxy
is resulting in this issue. The keystone middleware is widely used by
other openstack components. One possibility is that this kind of issue
is related to the usage of deprecated web frameworks in the proxy
implementation [1].

[1] https://eventlet.readthedocs.io/en/latest/index.html

Test Plan:

Nova action repetitions:
- PASS: Build and install openstack tarball on a duplex system
- PASS: Setup nova-api-proxy chart to use the image including the change
at https://review.opendev.org/c/starlingx/nfv/+/932568
- PASS: Perform 30+ repetitions of instance actions (suspend/resume)

Openstack sanity pipeline (automated tests) on duplex system:
- PASS: Live migrate VM
- PASS: Cold migrate VM
- PASS: Pause/unpause VM
- PASS: Suspend/resume VM

Closes-Bug: #2083934
Closes-Bug: #2083608

Depends-On: https://review.opendev.org/c/starlingx/nfv/+/932568

Change-Id: I21d02acc8fb669b657c84b09ee2e23b33ec5626d
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>
Co-authored-by: José Pires <joseclaudio.paespires@windriver.com>
Co-authored-by: Romulo Leite <romulo.leite@windriver.com>