Fix NFV-VIM error when listing volumes
It was noticed that the periodic NFV-VIM check on OpenStack volumes is
always failing and logging errors due to an invalid Cinder URL. The
reason for this behavior is that the function retrieving the Cinder
service name and type calls a method that always returns the v2 pattern
as the Cinder version, while the version configured in the OpenStack
endpoint list is v3, which should be used. The Cinder v2 configuration
is deprecated, so the solution for this issue is to update it to v3. To
achieve this, two new functions that receive the Cinder version were
created, adding the capability to send the version parameter as 'v3'.

TEST PLAN:

PASS - build-pkgs -c -p python3-k8sapp-openstack
PASS - build-pkgs --all
PASS - AIO-SX fresh install
PASS - Upload/Apply stx-openstack
PASS - 'openstack endpoint list' showing cinderv3

Partial-Bug: 2084683

Change-Id: I2c6717c200ebcdf7dc5ab6f2d2c5ee05ed5f7a8a
Signed-off-by: marantes <murillo.arantes@windriver.com>