Fix missing TLS certificates on host swact
This patch introduces an important improvement for managing OpenStack’s
HTTPS configuration, specifically targeting the reliability of
certificate handling across controller swacts and related system actions.
The main goal here is to make sure HTTPS remains consistently enabled and
that certificate files are always available and correctly synchronized,
regardless of changes in the active controller.

Previously, HTTPS could be disabled after a swact if the certificate
files weren't synchronized between controllers. This was due to a lack
of robust checks and syncing mechanisms for the certificate files
themselves, leading to situations where a failover would result in
missing certificates on the new active controller.

To address this, the patch adds new functions and updates existing ones
to introduce a logical flow for verifying and syncing certificate files.
The new logic controls the source from which certificate values are read.
By default, the system either checks for certificate files in the system,
based on the user defined paths, or use the already existing TLS secrets
to get the values of the certificates.

The new logic allows the certificate files' content to be defined by
either a pre-existing secret (meaning that this is not the first apply),
or by reading the contents of a certificate file in the system,
meaning that this is the first apply, or that the
`forceReadCertificateFiles` value is set to `true`, which forces the
application to bypass these overrides and read directly from the
specified file paths. This option is useful when operators want to ensure
that certificates are always sourced from the local file system.

Along with these core changes, this patch also updates test cases to
reflect the new logic and includes helper functions to streamline the
certificate file reading and validation processes. By integrating these
functions, the patch ensures that HTTPS remains reliable and that
certificates are always synchronized across controllers, providing a more
robust and predictable HTTPS setup for OpenStack deployments.

Test Plan:

- PASS: Build python3-k8sapp-openstack, stx-openstack-helm-fluxcd
        packages and STX-Openstack tarball
- PASS: Instal STX-Openstack on an AIO-SX with HTTPS enabled
- PASS: Intall STX-Openstack on a system with 2 controllers
- PASS: Enable HTTPS by installing the certificates via Helm overrides.
- PASS: After HTTPS is enabled, swact the active controller and verify
        that the certificate files were synced on both controllers.
- PASS: After the swact, verify that HTTPS is still enabled.
- PASS: Update certificate file, and, using the
        `forceReadCertificateFiles` set to true, reapply the
        application. Make sure the certificates were updated as well

Closes-Bug: 2085450

Change-Id: I6e18d4f1377732c89f67b39748c5cabb26ebcc1e
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>