Function to validate if NetApp is available
STX-O needs to support NetApp storage backend. A new function should be
created to validate if there is support for the NetApp storage system.
Whenever it is necessary for Cinder, Glance or Nova to check for NetApp
storage backend support, this function will be used.

This commit creates a new function that returns True if NetApp backend
is available and False if it is not.

Test Plan:
PASS - Validate True return when NetApp backend is available
PASS - Validate False return when NetApp backend is not available
PASS - Validate False if wrong namespace is used
PASS - Build / upload / apply STX-Openstack

Story: 2011281
Task: 51342

Change-Id: I1f3cce6ff21b3babd26cbf18648962da1b5eaab8
Signed-off-by: Johnny Chia <johnny.chialung@windriver.com>