Apply dynamic FQDN pattern in novncproxy endpoint
After [1] was introduced, it was noted that if the end user changed the
FQDN pattern ([2]), the novncproxy endpoint would not follow it. This
was caused because the function that sets the URL for this service was
not calling the function that uses this pattern, but instead was just
using a similar logic.

This change simply updates the `_get_novncproxy_base_url` function to
use the user-defined FQDN endpoint pattern instead of a hard-coded one.

[1] https://opendev.org/starlingx/openstack-armada-app/commit/0c02ebf1e1735944ddf22a76ce4119e84478632e
[2] https://opendev.org/starlingx/openstack-armada-app/src/branch/master/stx-openstack-helm-fluxcd/stx-openstack-helm-fluxcd/helm-charts/clients/values.yaml#L219

Test Plan:
PASS: Build / Upload / Apply STX-Openstack
PASS: Modify the `serviceEndpointPattern` value for a custom one
      ("{service_name}.{endpoint_domain}" for example)
PASS: After applying TLS, create a VM and access its console via
      Horizon

Closes-Bug: 2089165

Change-Id: I502ede740f834a7dfd7171bb205dc5e6e7b524a2
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>