Allow multiple volume backends in Cinder plugin
In preparation for the addition of the NetApp driver, some changes were
required in the Cinder Python plugin to allow for multiple backends to
be used at the same time. As STX-Openstack only supported Ceph so far,
this was not a requirements, so the plugin was written in a way that it
would either use Ceph or Ceph Rook as a backend. But now that we're
starting to support other backends as well, the logic for this plugin is
no longer valid.

Test Plan:
PASS: Build / Upload / Apply STX-Openstack
PASS: With Ceph configured as a backend, create a volume

Story: 2011281
Task: 51410

Change-Id: Id930bc29b5d4799fc65c5ce5943b21dec5d30d1f
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>
(cherry picked from commit 95e39ca669184487db9f95c4fe7612ba46d49c2c)