Enable NetApp as a volume backend for Cinder
This commit introduces support for using NetApp as a volume backend for
Cinder in STX-OpenStack Helm, adding compatibility with both NFS and
iSCSI protocols. The motivation for this enhancement is to provide
greater flexibility for deployments that leverage NetApp storage
solutions, allowing users to configure and manage their storage backends
more effectively. The implementation is designed to be modular, ensuring
that the NetApp backend is not enabled by default but can be activated
through user-defined overrides.

The code introduces configuration templates for the NetApp drivers in
the `values.yaml` file for the Cinder Helm chart. Two backends are
defined: `netapp-nfs`, which uses the NFS protocol, and `netapp-iscsi`,
which relies on iSCSI. These configurations include key parameters like
the NetApp storage protocol, server hostname, and authentication
credentials described in [1] and [2]. However, by default, these backends
are not included in the `enabled_backends` list, ensuring that they
remain inactive unless explicitly enabled by the user or the plugins.
The NetApp backends will be automatically added to the deployment of
Cinder if the plugin verifies that it is available in the cluster. If an
user has NetApp but don't want to use it in STX-Openstack, an override
can be defined for `conf.cinder.DEFAULT.enabled_backends`, which will
have a higher priority than the system overrides generated by the
plugin.

To support the NFS backend specifically, a new `nfs.shares` file was
introduced in the Cinder ConfigMap, providing a location to define NFS
shares for the NetApp driver. The Cinder volume deployment is updated to
mount this file, ensuring that it is accessible to the Cinder service.
The deployment templates now include logic to handle the mounting of
this file dynamically.

Utility functions are enhanced to improve the detection of available
NetApp backends. The function `check_netapp_backends`, which executes the
`tridentctl` command to determine if the required NetApp drivers
(`ontap-nas` for NFS or `ontap-san` for iSCSI) are available. This
function returns a dictionary indicating the availability of these
backends, which is then used to dynamically adjust the Cinder Helm
overrides. The existing `is_netapp_available` function now uses this
enhanced logic to provide a simplified check for whether any NetApp
backend is available.

In the Helm configuration logic, overrides for the NetApp backends were
added dynamically based on the results of these utility checks. If either
the NFS or iSCSI backend is available, the necessary parameters are
appended to the backend overrides, ensuring that they are included in the
final deployment configuration. This approach avoids hardcoding and
allows the deployment to adapt to the specific NetApp backends detected
in the environment.

The implementation ensures that NetApp backends are disabled by default,
preventing unintended configurations while enabling users to activate
them seamlessly when needed.

[1] https://netapp.github.io/openstack-deploy-ops-guide/juno/content/cinder.examples.cinder_conf.html
[2] https://wiki.openstack.org/wiki/How_to_deploy_cinder_with_NetApp

Test Plan:
PASS: Build / Upload / Apply STX-Openstack

Not using NetApp:
PASS: Verify that the NetApp backends are not enabled in cinder.conf
PASS: Create a volume and verify that it was stored in Ceph

Using NetApp NFS:
PASS: Apply Helm overrides with the configuration pointing to the
      NetApp cluster using an NFS backend
PASS: Verify that the netapp-nfs backend is enabled in cinder.conf
FAIL: Create a volume and verify that it was stored in NetApp by
      accessing the ONTAP dashboard

Using NetApp iSCSI:
PASS: Apply Helm overrides with the configuration pointing to the
      NetApp cluster using an iSCSI backend
PASS: Verify that the netapp-iscsi backend is enabled in cinder.conf

Failed test case is due to bug #2090845, which is already being worked
on.

Story: 2011281
Task: 51411
Task: 51412

Change-Id: I46cb68c5950e3343a3ed3a644981f558542e53d1
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>