Create Ingress-nginx FluxCD Manifest
The Openstack upstream community have deprecated its
Ingress Helm chart and begun to use the Helm chart from
Nginx.

The purpose of this task is to add a new FluxCD Manifest
for the Ingress-nginx Helm chart without removing the
current Ingress manifest. There will be a follow-up to
this review that will substitute the current ingress
manifest with the new one. Because of that, ingress-nginx
is not being added to the kustomization file, as it is
not to be deployed right now.

The plugin for this new Helm chart is also added to the
helm folder.

Test Plan:
PASS - Build stx-openstack-helm-fluxcd and STX-O
PASS - Ingress-nginx Helm chart appears in the build
PASS - Update and apply STX-O
PASS - Ingress-nginx is present but not deployed

Story: 2011303
Task: 51430

Change-Id: Iaf3cb33724871141abb5f8334b5043d3b823041b
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>