Upversion python-openstackclient to v6.6.0-5
As part of the STX-Openstack upversion to CARACAL, the python-
openstackclient is being upversioned to version 6.6.0-5 [1], which is
the latest supported on CARACAL [2] and Debian.

The commit hash of the used repo is:
e90cc3256ba56481729ceacebe5563f20636d1dd

All patches were updated in order to make the package buildable in the
new version.

[1] https://salsa.debian.org/openstack-team/clients/python-openstackclient/-/tree/debian/6.6.0-5?ref_type=tags
[2] https://releases.openstack.org/caracal/index.html#caracal-python-openstackclient

Story: 2011303
Task: 51497

Test Plan:
PASS: Build the python-openstackclient package
PASS: Build the stx-openstackclients image
PASS: Tested custom image in a running stx-openstack
PASS: Check version used in system on test plan

Depends-on: https://review.opendev.org/c/starlingx/root/+/938292

Change-Id: I94117a3d60554043c12c6b8ab5908c6e3238dc75
Signed-off-by: Jose Claudio <joseclaudio.paespires@windriver.com>